backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
akshare
httpx==0.26.0
openai==1.12.0
apscheduler==3.10.4
aiosqlite==0.19.0
greenlet==3.0.3
asyncpg==0.29.0
redis==5.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
